new SearchForm screen

+++ aviasalestemplate/src/main/java/ru/aviasales/template/utils/CurrencyUtils.java
@@ -17,7 +17,7 @@ public class CurrencyUtils {
 		return AviasalesSDK.getInstance().getSearchData().getCurrencyRates();
 	}
 
-	public static long getPriceInAppCurrency(int priceInDefaultCur, String appCurCode, Map<String, Double> currencies) {
+	public static long getPriceInAppCurrency(long priceInDefaultCur, String appCurCode, Map<String, Double> currencies) {
 		if (currencies == null) return 0;
 		long price = priceInDefaultCur;
 		if (!appCurCode.equalsIgnoreCase(Defined.RESPONSE_DEFAULT_CURRENCY)) {
@@ -39,11 +39,11 @@ public class CurrencyUtils {
 
 
 	public static List<Currency> getCurrenciesList() {
-		Map<String,String> currencyCodes = Defined.getCurrenciesArray();
+		Map<String, String> currencyCodes = Defined.getCurrenciesArray();
 
 		List<Currency> currencies = new ArrayList<Currency>();
 
-		for (Map.Entry<String,String> currencyEntry : currencyCodes.entrySet()) {
+		for (Map.Entry<String, String> currencyEntry : currencyCodes.entrySet()) {
 
 			Currency currency = new Currency();
 			currency.setCode(currencyEntry.getKey());
